# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Electronic Document Management System API - Backend service for document storage, search, and management\nSupports document ingestion, full-text search, word cloud generation, and file system browsing",
        "title": "godocs Backend API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/drummonds/godocs"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/tags": {
            "get": {
                "description": "Retrieve all tags in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tags"
                ],
                "summary": "Get all tags",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.Tag"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new tag",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tags"
                ],
                "summary": "Create a new tag",
                "parameters": [
                    {
                        "description": "Tag object",
                        "name": "tag",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.Tag"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/database.Tag"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/document": {
            "delete": {
                "description": "Deletes a document or folder from the system, including database entry and physical file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Delete a file or folder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document ULID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "File path relative to document root",
                        "name": "path",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Document Deleted\" or \"Folder Deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "File not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/jobs": {
            "get": {
                "description": "Retrieve a list of recent jobs with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Get recent jobs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of jobs to return (default: 20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination (default: 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of jobs",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.Job"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/search": {
            "get": {
                "description": "Search all documents using PostgreSQL full-text search",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search"
                ],
                "summary": "Search documents",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search term",
                        "name": "term",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results",
                        "schema": {
                            "$ref": "#/definitions/engine.fullFileSystem"
                        }
                    },
                    "204": {
                        "description": "No results found"
                    },
                    "404": {
                        "description": "Empty search term",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "database.Job": {
            "type": "object",
            "properties": {
                "completedAt": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "currentStep": {
                    "description": "Human-readable current step",
                    "type": "string"
                },
                "error": {
                    "description": "Error message if failed",
                    "type": "string"
                },
                "id": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "message": {
                    "description": "Status message",
                    "type": "string"
                },
                "progress": {
                    "description": "0-100",
                    "type": "integer"
                },
                "result": {
                    "description": "JSON result data",
                    "type": "string"
                },
                "startedAt": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/database.JobStatus"
                },
                "totalSteps": {
                    "description": "Total number of steps",
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/database.JobType"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "database.JobStatus": {
            "type": "string",
            "enum": [
                "pending",
                "running",
                "completed",
                "failed",
                "cancelled"
            ],
            "x-enum-varnames": [
                "JobStatusPending",
                "JobStatusRunning",
                "JobStatusCompleted",
                "JobStatusFailed",
                "JobStatusCancelled"
            ]
        },
        "database.JobType": {
            "type": "string",
            "enum": [
                "ingestion",
                "cleanup",
                "wordcloud",
                "search_reindex"
            ],
            "x-enum-varnames": [
                "JobTypeIngestion",
                "JobTypeCleanup",
                "JobTypeWordCloud",
                "JobTypeSearchReindex"
            ]
        },
...
        "database.Tag": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "sort_order": {
                    "type": "integer"
                },
                "tag_group": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "engine.fileTreeStruct": {
            "type": "object",
            "properties": {
                "childrenIDs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "fileURL": {
                    "type": "string"
                },
                "fullPath": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isDir": {
                    "type": "boolean"
                },
                "modDate": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "openable": {
                    "type": "boolean"
                },
                "parentID": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "thumbnailURL": {
                    "type": "string"
                },
                "ulid": {
                    "type": "string"
                }
            }
        },
        "engine.fullFileSystem": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "fileSystem": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/engine.fileTreeStruct"
                    }
                },
                "warnings": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    },
...
}